% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\name{adding_ai_footnotes}
\alias{adding_ai_footnotes}
\title{Update footnote with AI response}
\usage{
adding_ai_footnotes(outputs, prompt_list, platform, base_url, api_key, model)
}
\arguments{
\item{outputs}{Output objects}

\item{prompt_list}{List of prompt}

\item{platform}{Platform provider}

\item{base_url}{Base url}

\item{api_key}{API key}

\item{model}{Model of choice}
}
\description{
Update footnote with AI response
}
