% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoslideR_error.R
\name{autoslider_error}
\alias{autoslider_error}
\title{autoslider_error class}
\usage{
autoslider_error(x, spec, step)
}
\arguments{
\item{x}{character scaler}

\item{spec}{spec should be a list containing "program" and "suffix"}

\item{step}{step is a character indicating in which step the pipeline encounter error}
}
\value{
autoslider_error object
}
\description{
autoslider_error class
}
\details{
this function is used to create autoslider_error object.
this function is for internal use only to create the autoslider_error object.
It enable us for further functionalities, like providing help on easy debugging,
e.g. if the error is inside the user function, provide the call and let the user
run the code outside the pipeline.
}
