% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_slides.R
\name{center_figure_loc}
\alias{center_figure_loc}
\title{Create location container to center the figure, based on ppt size and
user specified figure size}
\usage{
center_figure_loc(fig_width, fig_height, ppt_width, ppt_height)
}
\arguments{
\item{fig_width}{Figure width}

\item{fig_height}{Figure height}

\item{ppt_width}{Slide width}

\item{ppt_height}{Slide height}
}
\value{
Location for a placeholder from scratch
}
\description{
Create location container to center the figure, based on ppt size and
user specified figure size
}
