% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_slides.R
\name{center_table_loc}
\alias{center_table_loc}
\title{create location container to center the table}
\usage{
center_table_loc(ft, ppt_width, ppt_height)
}
\arguments{
\item{ft}{Flextable object}

\item{ppt_width}{Powerpoint width}

\item{ppt_height}{Powerpoint height}
}
\value{
Location for a placeholder
}
\description{
create location container to center the table
}
