% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_and_set_cutoff}
\alias{check_and_set_cutoff}
\title{Assert function to check the cutoff}
\usage{
check_and_set_cutoff(data, cutoff)
}
\arguments{
\item{data}{dataframe}

\item{cutoff}{cutoff threshold}
}
\description{
Assert function to check the cutoff
}
