% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate,VTableTree-method}
\alias{decorate,VTableTree-method}
\title{Decorate TableTree}
\usage{
\S4method{decorate}{VTableTree}(x, titles = "", footnotes = "", paper = "P8", for_test = FALSE, ...)
}
\arguments{
\item{x}{A VTableTree object representing the data to be decorated.}

\item{titles}{Title to be added to the table.}

\item{footnotes}{Footnote to be added to the table}

\item{paper}{Orientation and font size as string, e.g. "P8"; "L11"}

\item{for_test}{`logic` CICD parameter}

\item{...}{Additional arguments passed to the decoration function.}
}
\description{
Decorate TableTree
}
