% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate.autoslider_error}
\alias{decorate.autoslider_error}
\title{decorate method for autoslider_error class}
\usage{
\method{decorate}{autoslider_error}(x, ...)
}
\arguments{
\item{x}{object to decorate}

\item{...}{additional arguments. not used.}
}
\value{
No return value, called for side effects
}
\description{
decorate method for autoslider_error class
}
