% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate.default}
\alias{decorate.default}
\title{default method to decorate}
\usage{
decorate.default(x, ...)
}
\arguments{
\item{x}{object to decorate}

\item{...}{additional arguments. not used.}
}
\value{
No return value, called for side effects
}
\description{
default method to decorate
}
