% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate.list}
\alias{decorate.list}
\title{decorate list of grobs}
\usage{
\method{decorate}{list}(
  x,
  titles,
  footnotes,
  paper = "L11",
  for_test = FALSE,
  metadata = NULL,
  ...
)
}
\arguments{
\item{x}{object to decorate}

\item{titles}{graph titles}

\item{footnotes}{graph footnotes}

\item{paper}{paper size. default is "L11".}

\item{for_test}{`logic` CICD parameter}

\item{metadata}{Named `list` (or `NULL`) of token values used to substitute
`\{token\}` placeholders in `titles`/`footnotes`. See [apply_tokens()].}

\item{...}{additional arguments. not used}
}
\value{
No return value, called for side effects
}
\description{
decorate list of grobs
}
\details{
The paper default paper size, `L11`, indicate that the fontsize is 11.
The fontsize of the footnotes, is the fontsize of the titles minus 2.
}
