% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate.listing_df}
\alias{decorate.listing_df}
\title{decorate listing}
\usage{
\method{decorate}{listing_df}(
  x,
  titles = "",
  footnotes = "",
  paper = "P8",
  for_test = FALSE,
  metadata = NULL,
  ...
)
}
\arguments{
\item{x}{A listing_df object representing the data to be decorated.}

\item{titles}{Title to be added to the table.}

\item{footnotes}{Footnote to be added to the table}

\item{paper}{Orientation and font size as string, e.g. "P8"; "L11"}

\item{for_test}{`logic` CICD parameter}

\item{metadata}{Named `list` (or `NULL`) of token values used to substitute
`\{token\}` placeholders in `titles`/`footnotes`. See [apply_tokens()].}

\item{...}{Additional arguments. not used.}
}
\value{
No return value, called for side effects
}
\description{
decorate listing
}
