% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate}
\alias{decorate}
\title{decorate generic}
\usage{
decorate(x, ...)
}
\arguments{
\item{x}{object to decorate}

\item{...}{additional arguments passed to methods}
}
\description{
decorate generic
}
