% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate.tbl_roche_summary}
\alias{decorate.tbl_roche_summary}
\title{decorate method for tbl_roche_summary (NEST 2 gtsummary subclass)}
\usage{
\method{decorate}{tbl_roche_summary}(x, ...)
}
\arguments{
\item{x}{tbl_roche_summary object to decorate}

\item{...}{arguments passed to \code{decorate.gtsummary}}
}
\description{
Handles objects whose class vector is set to just \code{"tbl_roche_summary"}
(losing the full gtsummary hierarchy) as well as those with the full
\code{c("tbl_roche_summary", "tbl_summary", "gtsummary")} hierarchy.
}
