% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{eg_adae}
\alias{eg_adae}
\title{Cached ADAE}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 93 columns.
}
\usage{
data(eg_adae)
}
\description{
Cached ADAE data
}
\keyword{datasets}
