% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{eg_adeg}
\alias{eg_adeg}
\title{Cached ADEG}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13600 rows and 88 columns.
}
\usage{
data(eg_adeg)
}
\description{
Cached ADEG data
}
\keyword{datasets}
