% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{eg_adex}
\alias{eg_adex}
\title{Cached ADEX}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6400 rows and 79 columns.
}
\usage{
data(eg_adex)
}
\description{
Cached ADEX data
}
\keyword{datasets}
