% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{eg_adlb}
\alias{eg_adlb}
\title{Cached ADLB}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 102 columns.
}
\usage{
data(eg_adlb)
}
\description{
Cached
ADLB data
}
\keyword{datasets}
