% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{eg_adrs}
\alias{eg_adrs}
\title{Cached ADRS}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3200 rows and 65 columns.
}
\usage{
data(eg_adrs)
}
\description{
Cached ADRS data
}
\keyword{datasets}
