% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{eg_adsl}
\alias{eg_adsl}
\title{Cached ADSL}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 56 columns.
}
\usage{
data(eg_adsl)
}
\description{
Cached ADSL data
}
\keyword{datasets}
