% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{eg_adtr}
\alias{eg_adtr}
\title{Cached ADTR}
\format{
An object of class \code{data.frame} with 2800 rows and 76 columns.
}
\usage{
data(eg_adtr)
}
\description{
Cached ADTR data
}
\keyword{datasets}
