% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{eg_adtte}
\alias{eg_adtte}
\title{Cached ADTTE}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 67 columns.
}
\usage{
data(eg_adtte)
}
\description{
Cached ADTTE data
}
\keyword{datasets}
