% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{eg_advs}
\alias{eg_advs}
\title{Cached ADVS}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16800 rows and 87 columns.
}
\usage{
data(eg_advs)
}
\description{
Cached ADVS data
}
\keyword{datasets}
