% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastdocall.R
\name{fastDoCall}
\alias{fastDoCall}
\title{Does do.call quicker, and avoids issues with debug mode within do.call}
\usage{
fastDoCall(what, args, quote = FALSE, envir = parent.frame())
}
\arguments{
\item{what}{either a function or a non-empty character string naming the function to be called.}

\item{args}{a list of arguments to the function call. The names attribute of args gives the argument names.}

\item{quote}{a logical value indicating whether to quote the arguments.}

\item{envir}{an environment within which to evaluate the call. This will be most useful if what is a character
string and the arguments are symbols or quoted expressions.}
}
\description{
copied from ms showcase app
}
