% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_slides.R
\name{figure_to_slide}
\alias{figure_to_slide}
\title{Add figure to slides}
\usage{
figure_to_slide(
  ppt,
  content,
  decor = TRUE,
  fig_width,
  fig_height,
  figure_loc = ph_location_type("body"),
  editable = FALSE,
  ...
)
}
\arguments{
\item{ppt}{slide page}

\item{content}{content to be added}

\item{decor}{should decoration be added}

\item{fig_width}{user specified figure width}

\item{fig_height}{user specified figure height}

\item{figure_loc}{location of the figure. Defaults to `ph_location_type("body")`}

\item{...}{arguments passed to program}
}
\value{
slide with the added content
}
\description{
Add figure to slides
}
