% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{format_3d}
\alias{format_3d}
\title{Format of xx.xx (xx.xx, xx.xx)}
\usage{
format_3d(x, output)
}
\arguments{
\item{x}{input array}

\item{output}{output handle}
}
\description{
Format of xx.xx (xx.xx, xx.xx)
}
