% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_wrapper.r
\name{func_wrapper}
\alias{func_wrapper}
\title{function wrapper to pass filtered data}
\usage{
func_wrapper(func, datasets, spec, verbose = TRUE)
}
\arguments{
\item{func}{function name}

\item{datasets}{list of raw datasets}

\item{spec}{spec}

\item{verbose}{whether to show verbose information}
}
\value{
a wrapped function using filtered adam
}
\description{
function wrapper to pass filtered data
}
