% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_slides.R
\name{get_proper_title}
\alias{get_proper_title}
\title{Adjust title line break and font size}
\usage{
get_proper_title(title, max_char = 60, title_color = "#1C2B39")
}
\arguments{
\item{title}{Character string}

\item{max_char}{Integer specifying the maximum number of characters in one line}

\item{title_color}{Title color}
}
\description{
Adjust title line break and font size
}
