% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlg_template.R
\name{get_template_filepath}
\alias{get_template_filepath}
\title{Retrieve Template File Paths}
\usage{
get_template_filepath(package = "autoslider.core", full.names = FALSE)
}
\arguments{
\item{package}{A character string specifying the name of the package to search.}

\item{full.names}{If `TRUE`, returns the full path to each file.
If `FALSE`, returns only the file names.}
}
\value{
A character vector of template file names or paths, depending on `full.names`.
}
\description{
Retrieve Template File Paths
}
\keyword{internal}
