% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{if_display_pp}
\alias{if_display_pp}
\title{Checker of data frame display per patient (or not)}
\usage{
if_display_pp(pp, colname = FALSE)
}
\arguments{
\item{pp}{per patient}

\item{colname}{name of column}
}
\description{
Checker of data frame display per patient (or not)
}
