% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_vs_slide.R
\name{l_vs_slide}
\alias{l_vs_slide}
\title{Listing of Vital Signs: Safety-Evaluable Patients}
\usage{
l_vs_slide(adsl, advs, trt_var = "TRT01A")
}
\arguments{
\item{adsl}{A dataframe representing the `ADSL` data.
Used to propagate filters to `ADVS`}

\item{advs}{A dataframe representing the ADVS data.}

\item{trt_var}{Character scalar or `NULL`. `"TRT01A"` by default. Specifies which variable,
if any, should be included as the treatment arm.}
}
\value{
A listing object of vital signs of patients
}
\description{
This function generates a listing with vital signs of patients.
It creates various columns such as patient identifier, age/sex/race, visit,
and calculates the range of several vital signs. The output listing includes labels for each column and a main title.
}
\details{
The function first validates whether the required columns are present in the ADVS dataframe.
Then, various calculations and transformations are performed to create additional columns.
Finally, the data is formatted into a listing with appropriate column labels.
}
\examples{
require(dplyr)
require(rlistings)
adsl <- eg_adsl
advs <- eg_advs |> filter(SUBJID \%in\% c("id-236", "id-65", "id-93"))

# Default usage
l_vs_slide(adsl = adsl, advs = advs, trt_var = "TRT01A")

# Custom treatment variable
l_vs_slide(adsl = adsl, advs = advs, trt_var = NULL)

}
\seealso{
\href{https://insightsengineering.github.io/tlg-catalog/stable/listings/vital-signs/vsl01.html}{Vital signs Listing}
}
