% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlg_template.R
\name{list_all_templates}
\alias{list_all_templates}
\title{[EXPERIMENTAL] List All Available Templates}
\usage{
list_all_templates(package = "autoslider.core")
}
\arguments{
\item{package}{Which package to search for the template files. Defaults to "autoslider.core".}
}
\value{
A character vector of available template names in the specified package.
}
\description{
[EXPERIMENTAL] List All Available Templates
}
\examples{
list_all_templates(package = "autoslider.core")
}
