% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{lyt_to_side_by_side_two_data}
\alias{lyt_to_side_by_side_two_data}
\title{Build side by side layout by cbind}
\usage{
lyt_to_side_by_side_two_data(lyt, anl, alt_counts_df, side_by_side = NULL)
}
\arguments{
\item{lyt}{layout object}

\item{anl}{analysis data object}

\item{alt_counts_df}{alternative data frame for counts}

\item{side_by_side}{A logical value indicating whether to display the data side by side.}
}
\value{
An `rtables` layout
}
\description{
Build side by side layout by cbind
}
