% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mutate_actarm}
\alias{mutate_actarm}
\title{Refactor active arm}
\usage{
mutate_actarm(
  df,
  arm_var = "TRT01A",
  levels = c("PLACEBO + PACLITAXEL + CISPLATIN",
    "ATEZOLIZUMAB + TIRAGOLUMAB + PACLITAXEL + CISPLATIN"),
  labels = c("Pbo+Pbo+PC", "Tira+Atezo+PC")
)
}
\arguments{
\item{df}{Input dataframe}

\item{arm_var}{Arm variable}

\item{levels}{factor levels}

\item{labels}{factor labels}
}
\value{
Dataframe with re-level and re-labelled arm variable.
}
\description{
Refactor active arm
}
