% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{na_replace}
\alias{na_replace}
\title{Replace NAs to NA}
\usage{
na_replace(table_df)
}
\arguments{
\item{table_df}{Table dataframe}
}
\value{
Input dataframe with both column replaced to NA
}
\description{
Replace NAs to NA
}
