% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{new_round}
\alias{new_round}
\title{Founding method}
\usage{
new_round(x, digits = 1)
}
\arguments{
\item{x}{number need to be rounded}

\item{digits}{number of digits}
}
\value{
rounded value
}
\description{
Founding method
}
