% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_slides.R
\name{ph_with_img}
\alias{ph_with_img}
\title{Placeholder for ph_with_img}
\usage{
ph_with_img(ppt, figure, fig_width, fig_height, figure_loc)
}
\arguments{
\item{ppt}{power point file}

\item{figure}{image object}

\item{fig_width}{width of figure}

\item{fig_height}{height of figure}

\item{figure_loc}{location of figure}
}
\value{
Location for a placeholder
}
\description{
Placeholder for ph_with_img
}
