% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{preprocess_t_dd}
\alias{preprocess_t_dd}
\title{Preprocess t_dd function}
\usage{
preprocess_t_dd(
  df,
  levels = c("PROGRESSIVE DISEASE", "ADVERSE EVENT", "OTHER", "<Missing>"),
  labels = c("Progressive Disease", "Adverse Events", "Other", "<Missing>")
)
}
\arguments{
\item{df}{Input dataframe}

\item{levels}{factor levels}

\item{labels}{factor labels}
}
\description{
Preprocess t_dd function
}
