% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{preprocess_t_ds}
\alias{preprocess_t_ds}
\title{Preprocess t_ds function}
\usage{
preprocess_t_ds(
  df,
  levels = c("Alive: On Treatment", "Alive: In Follow-up", "<Missing>"),
  labels = c("Alive: On Treatment", "Alive: In Follow-up", "<Missing>")
)
}
\arguments{
\item{df}{Input dataframe}

\item{levels}{factor levels}

\item{labels}{factor labels}
}
\description{
Preprocess t_ds function
}
