% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{print.decoratedGrob}
\alias{print.decoratedGrob}
\title{Print decorated grob}
\usage{
\method{print}{decoratedGrob}(x, ...)
}
\arguments{
\item{x}{An object of class \code{decoratedGrob}}

\item{...}{not used.}
}
\value{
No return value, called for side effects
}
\description{
Print decorated grob
}
