% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{print.decoratedGrobSet}
\alias{print.decoratedGrobSet}
\title{Print decorated grob set}
\usage{
\method{print}{decoratedGrobSet}(x, ...)
}
\arguments{
\item{x}{An object of class `decoratedGrobSet`}

\item{...}{not used.}
}
\description{
Print decorated grob set
}
