% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{s_surv_time_1}
\alias{s_surv_time_1}
\title{survival time afun}
\usage{
s_surv_time_1(df, .var, is_event, control = control_surv_time())
}
\arguments{
\item{df}{data}

\item{.var}{variable of interest}

\item{is_event}{vector indicating event}

\item{control}{\code{control_surv_time()} by default}
}
\value{
A function suitable for use in rtables::analyze() with element selection,
reformatting, and relabeling performed automatically.
}
\description{
survival time afun
}
