% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_slides.R
\name{slides_preview}
\alias{slides_preview}
\title{Generate flextable for preview first page}
\usage{
slides_preview(x)
}
\arguments{
\item{x}{rtables or data.frame}
}
\value{
A flextable or a ggplot object depending to the input.
}
\description{
Generate flextable for preview first page
}
\examples{
# Example 1. preview table
library(dplyr)
adsl <- eg_adsl
t_dm_slide(adsl, "TRT01P", c("SEX", "AGE")) |> slides_preview()
}
