% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ae_pt_diff_slide_to_df.R
\name{t_ae_pt_diff_slide_to_df}
\alias{t_ae_pt_diff_slide_to_df}
\title{autoslider: t_ae_pt_diff_slide to dataframe}
\usage{
t_ae_pt_diff_slide_to_df(t_ae_pt_diff_slide)
}
\arguments{
\item{t_ae_pt_diff_slide}{an output from the autoslider function `t_ae_pt_diff_slide`
See \link[autoslideR]{t_ae_pt_diff_slide}}
}
\value{
dataframe object
}
\description{
autoslider: t_ae_pt_diff_slide to dataframe
}
\examples{
library(dplyr)
adsl <- eg_adsl \%>\%
  mutate(TRT01A = factor(TRT01A, levels = c("A: Drug X", "B: Placebo")))
adae <- eg_adae \%>\%
  mutate(TRT01A = factor(TRT01A, levels = c("A: Drug X", "B: Placebo")))
adae$ATOXGR <- adae$AETOXGR
my_tbl <- t_ae_pt_diff_slide(adsl, adae, "TRT01A", 2)

t_ae_pt_diff_slide_to_df(my_tbl)
}
