% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ft_funs.R
\name{to_flextable.Ddataframe}
\alias{to_flextable.Ddataframe}
\title{To flextable}
\usage{
\method{to_flextable}{Ddataframe}(x, lpp, table_format = table_format, ...)

\method{to_flextable}{Ddataframe}(x, lpp, table_format = table_format, ...)
}
\arguments{
\item{x}{dataframe}

\item{lpp}{\{lpp\} from \{paginate_table\}. numeric. Maximum lines per page}

\item{table_format}{Table format}

\item{...}{arguments passed to program}
}
\description{
Convert the dataframe into flextable, and merge the cells
that have colspan > 1. align the columns to the middle, and the row.names to
the left. indent the row.names by 10 times indention.
}
\details{
convert the dataframe object into flextable, and merge the cells
that have colspan > 1. align the columns to the middle, and the row.names to
the left. indent the row.names by 10 times indention. titles are added in headerlines,
footnotes are added in footer lines,
The width of the columns are aligned based on autofit() of officer function.
For paginated table, the width of the 1st column are set as the widest 1st column among paginated tables
}
