% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ft_funs.R
\name{to_flextable.dVTableTree}
\alias{to_flextable.dVTableTree}
\title{To flextable}
\usage{
\method{to_flextable}{dVTableTree}(x, lpp, cpp, ...)
}
\arguments{
\item{x}{decorated rtable(dVTableTree) object}

\item{lpp}{\{lpp\} from \link[rtables]{paginate_table}. numeric. Maximum lines per page}

\item{...}{argument parameters}
}
\description{
To flextable
}
\details{
convert the VTableTree object into flextable, and merge the cells
that have colspan > 1. align the columns to the middle, and the row.names to
the left. indent the row.names by 10 times indention. titles are added in headerlines,
footnotes are added in footer lines,
The width of the columns are aligned based on autofit() of officer function.
For paginated table, the width of the 1st column are set as the widest 1st column among paginated tables
}
