% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ft_funs.R
\name{to_flextable.gtsummary}
\alias{to_flextable.gtsummary}
\title{convert gtsummary to flextable}
\usage{
\method{to_flextable}{gtsummary}(x, table_format = autoslider_format, ...)
}
\arguments{
\item{x}{gtsummary object}

\item{table_format}{Function applied to the flextable for styling}

\item{...}{additional arguments, not used}
}
\description{
convert gtsummary to flextable
}
