% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ft_funs.R
\name{to_flextable.tbl_roche_summary}
\alias{to_flextable.tbl_roche_summary}
\title{convert tbl_roche_summary (NEST 2 gtsummary subclass) to flextable}
\usage{
\method{to_flextable}{tbl_roche_summary}(x, ...)
}
\arguments{
\item{x}{tbl_roche_summary object}

\item{...}{arguments passed to \code{to_flextable.gtsummary}}
}
\description{
convert tbl_roche_summary (NEST 2 gtsummary subclass) to flextable
}
