% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ft_funs.R
\name{to_flextable.VTableTree}
\alias{to_flextable.VTableTree}
\title{Covert rtables object to flextable}
\usage{
\method{to_flextable}{VTableTree}(x, table_format = orange_format, ...)
}
\arguments{
\item{x}{rtable(VTableTree) object}

\item{table_format}{a function that decorate a flextable and return a flextable}
}
\description{
Covert rtables object to flextable
}
