% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{to_vector}
\alias{to_vector}
\title{Convert list of numbers to vectors}
\usage{
to_vector(num_list)
}
\arguments{
\item{num_list}{list of numbers}
}
\value{
No return value, called for side effects
}
\description{
Convert list of numbers to vectors
}
