% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{trim_perc}
\alias{trim_perc}
\title{Format of xx.xx (xx.x)}
\usage{
trim_perc(x, output)
}
\arguments{
\item{x}{input array}

\item{output}{output handle}
}
\value{
formatted values
}
\description{
Format of xx.xx (xx.x)
}
