% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlg_template.R
\name{use_template}
\alias{use_template}
\title{[EXPERIMENTAL] Create new output function based on a template.}
\usage{
use_template(
  template = "t_dm_slide",
  function_name = "default",
  save_path = "./programs/R",
  overwrite = FALSE,
  open = interactive(),
  package = "autoslider.core"
)
}
\arguments{
\item{template}{Must be one of `list_all_templates(package = "autoslider.core")`.}

\item{function_name}{Name of the output function you want to create. Defaults to "default".}

\item{save_path}{Path to save the function. Defaults to "./programs/R".}

\item{overwrite}{Whether to overwrite an existing file.}

\item{open}{Whether to open the file after creation.}

\item{package}{Which package to search for the template file. Defaults to "autoslider.core".}
}
\value{
No return value. Called for side effects (writes a file).
}
\description{
We have separate templates for listings, tables, and graphs.
There is also a template to set up the `run_all` script.
}
\details{
Use `list_all_templates(package = "autoslider.core")` to discover which templates are available.
}
\examples{
if (interactive()) {
  use_template("t_dm_slide", function_name = "my_table", package = "autoslider.core")
}
}
