% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{autoslider.core-package}
\alias{autoslider.core}
\alias{autoslider.core-package}
\title{autoslider.core Package}
\description{
Slides automation with R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dss-rac.pages.roche.com/autoslideR/}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{joe.zhu@roche.com}

Authors:
\itemize{
  \item Heng Wang \email{wangh107@gene.com}
  \item Yinqi Zhao \email{zhaoy143@gene.com}
  \item Bo Ci \email{ci.bo@gene.com}
  \item Liming Li \email{liming.li@roche.com}
  \item Chenkai Lv \email{chenkai.lv@gene.com}
}

Other contributors:
\itemize{
  \item Laura Wang \email{wangl330@gene.com} [contributor]
  \item Xiaoli Duan \email{xiaoli.duan@roche.com} [contributor]
  \item Stefan Pascal Thoma \email{stefan.thoma@roche.com} [contributor]
  \item Miles Almond \email{miles.almond@roche.com} [contributor]
  \item Mahdi About \email{mahdi.about@roche.com} [contributor]
  \item Kai Lim \email{kai.lim@roche.com} [contributor]
  \item Nolan Steed \email{nolan.steed@roche.com} [contributor]
  \item Daoling Pang \email{daolin.pang@roche.com} [contributor]
  \item Elisabeth Deutschmann \email{elisabeth.deutschmann@roche.com} [contributor]
}

}
