% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{autoslider.core-package}
\alias{autoslider.core}
\alias{autoslider.core-package}
\title{autoslider.core Package}
\description{
The normal process of creating clinical study slides is that BioStat manually type in the numbers from outputs and a separate BioStat to QC the typed in numbers. This process is time consuming, resource intensive, and error prone. Automatic slide generation is a solution to address these issues. It reduces the amount of work and the required time when creating slides, and reduces the risk of errors from manually typing or copying numbers from the output to slides. It also helps users to avoid unnecessary stress when creating large amounts of slide decks in a short time window
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/autoslider.core}
  \item Report bugs at \url{https://github.com/insightsengineering/autoslider.core/issues}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{joe.zhu@roche.com}

Authors:
\itemize{
  \item Heng Wang
  \item Yinqi Zhao
  \item Bo Ci
  \item Liming Li
  \item Xiaoli Duan
  \item Stefan Pascal Thoma
  \item Miles Almond
  \item Chenkai Lv
}

Other contributors:
\itemize{
  \item Laura Wang [contributor]
  \item Thomas Neitmann [contributor]
  \item Mahdi About [contributor]
  \item Kai Lim [contributor]
  \item Nolan Steed [contributor]
  \item Daoling Pang [contributor]
  \item Elisabeth Deutschmann [contributor]
}

}
