% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{autoslider.core-package}
\alias{autoslider.core}
\alias{autoslider.core-package}
\title{autoslider.core Package}
\description{
Clinical TLG slide automation empowered by `rtables`, `rlistings` and `ggplot2`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/autoslider.core}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{joe.zhu@roche.com}

Authors:
\itemize{
  \item Heng Wang
  \item Yinqi Zhao
  \item Bo Ci
  \item Liming Li
  \item Xiaoli Duan
  \item Stefan Pascal Thoma
  \item Miles Almond
  \item Chenkai Lv
}

Other contributors:
\itemize{
  \item Laura Wang [contributor]
  \item Thomas Neitmann [contributor]
  \item Mahdi About [contributor]
  \item Kai Lim [contributor]
  \item Nolan Steed [contributor]
  \item Daoling Pang [contributor]
  \item Elisabeth Deutschmann [contributor]
}

}
