% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate.ggplot}
\alias{decorate.ggplot}
\title{Decorate ggplot object}
\usage{
\method{decorate}{ggplot}(x, titles = "", footnotes = "", paper = "L11", for_test = FALSE, ...)
}
\arguments{
\item{x}{An object to decorate}

\item{titles}{Plot titles}

\item{footnotes}{Plot footnotes}

\item{paper}{Paper size, by default "L11"}

\item{for_test}{`logic` CICD parameter}

\item{...}{additional arguments. not used.}
}
\value{
No return value, called for side effects
}
\description{
Decorate ggplot object
}
\details{
The paper default paper size, `L11`, indicate that the fontsize is 11.
The fontsize of the footnotes, is the fontsize of the titles minus 2.
}
