% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate.grob}
\alias{decorate.grob}
\title{decorate grob}
\usage{
\method{decorate}{grob}(x, titles = "", footnotes = "", paper = "L11", for_test = FALSE, ...)
}
\arguments{
\item{x}{object to decorate}

\item{titles}{graph titles}

\item{footnotes}{graph footnotes}

\item{paper}{paper size. default is "L8".}

\item{for_test}{`logic` CICD parameter}

\item{...}{Additional arguments. not used.}
}
\value{
No return value, called for side effects
}
\description{
decorate grob
}
\details{
The paper default paper size, `L11`, indicate that the fontsize is 11.
The fontsize of the footnotes, is the fontsize of the titles minus 2.
}
