% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate}
\alias{decorate}
\title{generic function decorate}
\usage{
decorate(x, ...)

decorate(x, ...)
}
\arguments{
\item{x}{object to decorate}

\item{...}{additional arguments passed to methods}
}
\description{
generic function decorate

s3 method for decorate
}
