% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_mean_slides.R
\name{g_vs_slide}
\alias{g_vs_slide}
\title{Plot mean values of VS}
\usage{
g_vs_slide(
  adsl,
  advs,
  arm = "TRT01P",
  paramcd = "PARAM",
  subtitle = "Plot of Mean and 95\% Confidence Limits by Visit.",
  ...
)
}
\arguments{
\item{adsl}{ADSL data}

\item{advs}{ADVS data}

\item{arm}{`"TRT01P"` by default}

\item{paramcd}{Which variable to use for plotting. By default `"PARAM"`}

\item{subtitle}{character scalar forwarded to g_lineplot}

\item{...}{|
Gets forwarded to `tern::g_lineplot()`.
This lets you specify additional arguments to `tern::g_lineplot()`}
}
\description{
Wrapper for `g_mean_general()`.
Requires filtering of the datasets (e.g. using SUFFIX in spec.yml)
}
\examples{
library(dplyr)
advs_filtered <- eg_advs \%>\% filter(
  PARAMCD == "SYSBP"
)

plot_vs <- g_vs_slide(
  adsl = eg_adsl,
  advs = advs_filtered,
  paramcd = "PARAM",
  subtitle_add_unit = FALSE
)
plot_vs
# you want x-axis tilted labels? No problem:
plot_vs +
  ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 45, hjust = 1)) \%>\%
  generate_slides(paste0(tempdir(), "/g_vs.pptx"))
}
\author{
Stefan Thoma (`thomas7`)
}
