% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gen_notes}
\alias{gen_notes}
\title{General notes}
\usage{
gen_notes()
}
\description{
General notes
}
\note{
* Default arm variables are set to `"TRT01A"` for safety output, and
 `"TRT01P"` for efficacy output
}
