% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlg_template.R
\name{list_all_templates}
\alias{list_all_templates}
\title{[EXPERIMENTAL] List All Available Templates}
\usage{
list_all_templates(package = "autoslider.core")
}
\arguments{
  \item{package}{A character string specifying the package to search for template files.
  Defaults to \code{"autoslider.core"}.}
}
\value{
A character vector of all available template names in the specified package.
}
\description{
Return the names of all available template files (e.g., for tables, listings, and graphs) provided by a package.
}
\details{
This function searches for template files (those beginning with \code{t_}, \code{g_}, or \code{l_}) in the \code{inst/templates} folder of the specified package.
}
\examples{
list_all_templates(package = "autoslider.core")
}
\concept{utils_examples}
\keyword{utilities}
