% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_template.R
\name{list_all_templates}
\alias{list_all_templates}
\title{[EXPERIMENTAL] List All Available Templates}
\usage{
list_all_templates()
}
\value{
A `character` vector of all available templates
}
\description{
[EXPERIMENTAL] List All Available Templates
}
\examples{
list_all_templates()
}
\concept{utils_examples}
\keyword{utils_examples}
