% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{print.decoratedGrobSet}
\alias{print.decoratedGrobSet}
\title{Print decorated grob set}
\usage{
\method{print}{decoratedGrobSet}(x, ...)
}
\arguments{
\item{x}{An object of class \code{decoratedGrobSet}}

\item{...}{not used.}
}
\value{
No return value, called for side effects
}
\description{
Print decorated grob set
}
