% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_output.R
\name{save_output}
\alias{save_output}
\alias{save_output.autoslider_error}
\alias{save_output,dVTableTree-method}
\alias{save_output,}
\alias{dVTableTree,}
\alias{dVTableTree-method}
\alias{save_output.decoratedGrob}
\alias{save_output.decoratedGrobSet}
\alias{save_output.dlisting}
\title{Save an Output}
\usage{
save_output(output, file_name, save_rds = TRUE)

save_output(output, file_name, save_rds = TRUE)

save_output.autoslider_error(output, file_name, save_rds = TRUE)

\S4method{save_output}{dVTableTree}(output, file_name, save_rds = TRUE)

save_output.decoratedGrob(output, file_name, save_rds = TRUE)

save_output.decoratedGrobSet(output, file_name, save_rds = TRUE)

save_output.dlisting(output, file_name, save_rds = TRUE)
}
\arguments{
\item{output}{Output object, e.g. an `rtable` or `grob`}

\item{file_name}{Full path of the new file *excluding* the extension}

\item{save_rds}{Saved as an `.rds` files}
}
\value{
The input `object` invisibly
}
\description{
Save an Output
}
\details{
Tables are saved as RDS file
}
\examples{
library(dplyr)
adsl <- eg_adsl \%>\%
  mutate(TRT01A = factor(TRT01A, levels = c("A: Drug X", "B: Placebo")))
adae <- eg_adae \%>\%
  mutate(TRT01A = factor(TRT01A, levels = c("A: Drug X", "B: Placebo")))
adae$ATOXGR <- adae$AETOXGR
output_dir <- tempdir()
t_ae_pt_slide(adsl, adae, "TRT01A", 2) \%>\%
  decorate(
    title = "Serious Adverse Events, Safety-Evaluable Patients",
    footnote = ""
  ) \%>\%
  save_output(
    file_name = file.path(output_dir, "t_ae_SER_SE"),
    save_rds = TRUE
  )

}
