% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_output.R
\name{save_outputs}
\alias{save_outputs}
\title{Save a list of outputs}
\usage{
save_outputs(
  outputs,
  outfolder = file.path("output"),
  generic_suffix = NULL,
  save_rds = TRUE,
  verbose_level = 1
)
}
\arguments{
\item{outputs}{\code{list} of outputs as created by \code{generate_outputs}}

\item{outfolder}{Folder in which to store the `outputs``}

\item{generic_suffix}{generic suffix. must be length 1 character or NULL.}

\item{save_rds}{Should the input \code{outputs} be saved as \code{.rds} files in
in addition to \code{.out} or \code{.pdf} files? Defaults to \code{FALSE}.}

\item{verbose_level}{Level of verbose information displayed.
Default set to \code{1}.}
}
\value{
The input \code{object} invisibly
}
\description{
Save a list of outputs
}
\examples{
## As `save_outputs` is the last step in the pipeline we have to run
## the 'whole machinery' in order to show its functionality. Also take a look
## at the `AutoslideR-Demo` repo on code.roche.com.
library(dplyr, warn.conflicts = FALSE)

data <- list(
  adsl = eg_adsl,
  adae = eg_adae,
  adtte = eg_adtte
)

filters::load_filters(
  yaml_file = system.file("filters.yml", package = "autoslider.core"),
  overwrite = TRUE
)

## For this example the outputs will be saved in a temporary directory. In a
## production run this should be the reporting event's 'output' folder instead.
output_dir <- tempdir()

spec_file <- system.file("spec.yml", package = "autoslider.core")
read_spec(spec_file) \%>\%
  filter_spec(program == "t_dm_slide") \%>\%
  generate_outputs(datasets = data) \%>\%
  decorate_outputs() \%>\%
  save_outputs(outfolder = output_dir)

}
