% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_slides.R
\name{table_to_slide}
\alias{table_to_slide}
\title{Add decorated flextable to slides}
\usage{
table_to_slide(
  ppt,
  content,
  decor = TRUE,
  layout = "Title and Content",
  table_loc = ph_location_type("body"),
  usernotes = "",
  ...
)
}
\arguments{
\item{ppt}{Slide}

\item{content}{Content to be added}

\item{decor}{Should table be decorated}

\item{layout}{layout from theme}

\item{table_loc}{Table location}

\item{usernotes}{User notes}

\item{...}{additional arguments}
}
\value{
Slide with added content
}
\description{
Add decorated flextable to slides
}
