% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ft_funs.R
\name{to_flextable.data.frame}
\alias{to_flextable.data.frame}
\title{convert data.frame to flextable}
\usage{
\method{to_flextable}{data.frame}(
  x,
  col_width = NULL,
  table_format = orange_format,
  dose_template = FALSE,
  font_size = 9,
  ...
)
}
\description{
convert data.frame to flextable
}
