% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlg_template.R
\name{use_template}
\alias{use_template}
\title{[EXPERIMENTAL] Create new output function based on a template.}
\usage{
use_template(
  template = "t_dm_slide",
  function_name = "default",
  save_path = "./programs/R",
  overwrite = FALSE,
  open = interactive(),
  package = "autoslider.core"
)
}
\arguments{
  \item{template}{Name of the template to use. Must be one of those listed by
  \code{list_all_templates(package = "autoslider.core")}.}

  \item{function_name}{Name for the new template function being created.}

  \item{save_path}{Path to save the generated template file. If not provided, defaults to
  \code{"./programs/R/<function_name>.R"}.}

  \item{overwrite}{Whether to overwrite an existing file at \code{save_path}.}

  \item{open}{Logical. Whether to open the newly created file in an editor. Defaults to
  \code{interactive()}.}

  \item{package}{Which package to search for the template file. Default is
  \code{"autoslider.core"}.}
}
\value{
No return value. This function is called for its side effects (creating a file).
}
\description{
Create a new R function script based on a predefined template. Supports templates for tables, listings, and graphs.
}
\details{
You can use \code{list_all_templates(package = "<pkg>")} to discover available templates in a specific package.
}
\examples{
if (interactive()) {
  use_template("t_dm_slide", function_name = "example_function",
               package = "autoslider.core")
}
}
