% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlg_template.R
\name{use_template}
\alias{use_template}
\title{[EXPERIMENTAL] Create new output function based on a template.}
\usage{
use_template(
  template = "t_dm_slide",
  function_name = "default",
  save_path = "./programs/R",
  overwrite = FALSE,
  open = interactive(),
  package = "autoslider.core"
)
}
\arguments{
  \item{template}{Must be one of \code{list_all_templates(package = "package")}.}

  \item{function_name}{What output would you like to create?}

  \item{save_path}{Path to save the function, should include filename and extension.
  Defaults to \code{"./programs/R/<function_name>.R"}.}

  \item{overwrite}{Whether to overwrite an existing file named \code{save_path}.}

  \item{open}{Whether to open the script right away.}

  \item{package}{Which package to search for the template file. Default is \code{"autoslider.core"}.}
}
\value{
No return value. Called for side effects (creates a file).
}
\description{
Create a new R function file based on a predefined template for tables, listings, or graphs.
}
\details{
Use \code{list_all_templates(package = "example")} to discover which templates are available in the specified package.
}
\examples{
if (interactive()) {
  use_template("t_dm_slide", function_name = "example",
  package = "autoslider.core")
}
}
