% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_template.R
\name{use_template}
\alias{use_template}
\title{[EXPERIMENTAL] Create new output function based on a template.}
\usage{
use_template(
  template = "t_dm_slide",
  function_name = "default",
  save_path = NULL,
  overwrite = FALSE,
  open = interactive()
)
}
\arguments{
\item{template}{Must be one of `r list_all_templates()`}

\item{function_name}{What output would you like to create?}

\item{save_path}{Path to save the function, should include filename and extension.
Defaults to "./programs/**function_name**_slide.R"}

\item{overwrite}{Whether to overwrite an existing file named `save_path`.}

\item{open}{Whether to open the script right away.}
}
\value{
No return values, called for side effects
}
\description{
We have separate templates for listings, tables, and graphs.
There is also a template to set up the `run_all` script.
}
\details{
Use `list_all_templates()` to discover which templates are available.
}
\examples{
if (interactive()) {
  use_template("l_template_slide")
}
}
